transactions: separate each pre-signed tx in its own module
