+++ src/transactions/mod.rs
@@ -9,31 +9,35 @@ use crate::{error::*, scripts::*, txins::*, txouts::*};
 
 use miniscript::{
     bitcoin::{
-        consensus::encode::{Decodable, Encodable},
+        consensus::encode::Encodable,
         hash_types,
         hashes::Hash,
         secp256k1,
         util::{
-            bip143::SigHashCache,
-            bip32::ChildNumber,
-            psbt::{
-                Global as PsbtGlobal, Input as PsbtIn, Output as PsbtOut,
-                PartiallySignedTransaction as Psbt,
-            },
+            bip143::SigHashCache, bip32::ChildNumber, psbt::PartiallySignedTransaction as Psbt,
         },
         Address, Amount, Network, OutPoint, PublicKey as BitcoinPubKey, Script, SigHash,
         SigHashType, Transaction, Txid, Wtxid,
     },
-    BitcoinSig, DescriptorTrait,
+    BitcoinSig,
 };
 
-#[cfg(feature = "use-serde")]
-use {
-    serde::de::{self, Deserialize, Deserializer},
-    serde::ser::{Serialize, Serializer},
-};
+use std::fmt;
 
-use std::{collections::BTreeMap, convert::TryInto, fmt};
+#[macro_use]
+mod utils;
+
+mod cancel;
+mod emergency;
+mod spend;
+mod unvault;
+mod unvaultemergency;
+
+pub use cancel::CancelTransaction;
+pub use emergency::EmergencyTransaction;
+pub use spend::SpendTransaction;
+pub use unvault::UnvaultTransaction;
+pub use unvaultemergency::UnvaultEmergencyTransaction;
 
 /// The value of the CPFP output in the Unvault transaction.
 /// See [practical-revault](https://github.com/revault/practical-revault/blob/master/transactions.md#unvault_tx).
@@ -70,12 +74,11 @@ pub const MAX_STANDARD_TX_WEIGHT: u32 = 400_000;
 /// - Finalizer
 /// - Extractor and serializer
 pub trait RevaultTransaction: fmt::Debug + Clone + PartialEq {
-    // TODO: Eventually, we could not expose it and only have wrappers to access
-    // the PSBT informations
     /// Get the inner transaction
     fn inner_tx(&self) -> &Psbt;
 
-    // FIXME: don't expose this. Maybe a private trait?
+    // FIXME: how can we not expose this? This in theory breaks our internal assumptions as the
+    // caller could just put the inner PSBT in an insane state..
     /// Get the inner transaction
     fn inner_tx_mut(&mut self) -> &mut Psbt;
 
@@ -321,6 +324,7 @@ pub trait RevaultTransaction: fmt::Debug + Clone + PartialEq {
         Self::from_psbt_serialized(&base64::decode(&psbt_str)?)
     }
 
+    // TODO: just use rust-bitcoin's encode_hex
     /// Get the hexadecimal representation of the transaction as used by the bitcoind API.
     fn hex(&self) -> String {
         let buff = self.clone().into_bitcoin_serialized();
@@ -366,1096 +370,6 @@ pub trait RevaultTransaction: fmt::Debug + Clone + PartialEq {
     /// Get the inner unsigned transaction hash with witness data
     fn wtxid(&self) -> Wtxid {
         self.inner_tx().global.unsigned_tx.wtxid()
-    }
-}
-
-// Boilerplate for newtype declaration and small trait helpers implementation.
-macro_rules! impl_revault_transaction {
-    ( $transaction_name:ident, $doc_comment:meta ) => {
-        #[$doc_comment]
-        #[derive(Debug, Clone, PartialEq)]
-        pub struct $transaction_name(Psbt);
-
-        impl RevaultTransaction for $transaction_name {
-            fn inner_tx(&self) -> &Psbt {
-                &self.0
-            }
-
-            fn inner_tx_mut(&mut self) -> &mut Psbt {
-                &mut self.0
-            }
-
-            fn into_psbt(self) -> Psbt {
-                self.0
-            }
-
-            fn from_psbt_serialized(
-                raw_psbt: &[u8],
-            ) -> Result<Self, TransactionSerialisationError> {
-                $transaction_name::from_raw_psbt(raw_psbt)
-            }
-        }
-
-        #[cfg(feature = "use-serde")]
-        impl Serialize for $transaction_name {
-            fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-            where
-                S: Serializer,
-            {
-                if serializer.is_human_readable() {
-                    serializer.serialize_str(&self.as_psbt_string())
-                } else {
-                    serializer.serialize_bytes(&self.as_psbt_serialized())
-                }
-            }
-        }
-
-        #[cfg(feature = "use-serde")]
-        impl<'de> Deserialize<'de> for $transaction_name {
-            fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-            where
-                D: Deserializer<'de>,
-            {
-                if deserializer.is_human_readable() {
-                    $transaction_name::from_psbt_str(&String::deserialize(deserializer)?)
-                        .map_err(de::Error::custom)
-                } else {
-                    $transaction_name::from_psbt_serialized(&Vec::<u8>::deserialize(deserializer)?)
-                        .map_err(de::Error::custom)
-                }
-            }
-        }
-    };
-}
-
-// Sanity check a PSBT representing a RevaultTransaction, the part common to all transactions
-fn psbt_common_sanity_checks(psbt: Psbt) -> Result<Psbt, PsbtValidationError> {
-    let inner_tx = &psbt.global.unsigned_tx;
-
-    if inner_tx.version != TX_VERSION {
-        return Err(PsbtValidationError::InvalidTransactionVersion(
-            inner_tx.version,
-        ));
-    }
-
-    let input_count = inner_tx.input.len();
-    let psbt_input_count = psbt.inputs.len();
-    if input_count != psbt_input_count {
-        return Err(PsbtValidationError::InputCountMismatch(
-            input_count,
-            psbt_input_count,
-        ));
-    }
-
-    let output_count = inner_tx.output.len();
-    let psbt_output_count = psbt.outputs.len();
-    if output_count != psbt_output_count {
-        return Err(PsbtValidationError::OutputCountMismatch(
-            output_count,
-            psbt_output_count,
-        ));
-    }
-
-    // None: unknown, Some(true): an input was final, Some(false) an input was non-final
-    let mut is_final = None;
-    // Record the number of coins spent by the transaction
-    let mut value_in: u64 = 0;
-    for input in psbt.inputs.iter() {
-        // We restrict to native segwit, also for the external fee-bumping wallet.
-        if input.witness_utxo.is_none() {
-            return Err(PsbtValidationError::MissingWitnessUtxo(input.clone()));
-        }
-        let spk = &input.witness_utxo.as_ref().unwrap().script_pubkey;
-        if !(spk.is_v0_p2wsh() || spk.is_v0_p2wpkh()) {
-            return Err(PsbtValidationError::InvalidInputField(input.clone()));
-        }
-
-        if input.non_witness_utxo.is_some() {
-            return Err(PsbtValidationError::InvalidInputField(input.clone()));
-        }
-
-        if input.redeem_script.is_some() {
-            return Err(PsbtValidationError::InvalidInputField(input.clone()));
-        }
-
-        // Make sure it does not mix finalized and non-finalized inputs or final scripts
-        // and non-final scripts.
-        if input.final_script_witness.is_some() {
-            if is_final == Some(false) || input.witness_script.is_some() {
-                return Err(PsbtValidationError::PartiallyFinalized);
-            }
-            is_final = Some(true);
-        } else {
-            if is_final == Some(true) {
-                return Err(PsbtValidationError::PartiallyFinalized);
-            }
-            is_final = Some(false);
-        }
-
-        // If the witness script is provided, it must be a sane Miniscript
-        if let Some(ref script) = input.witness_script {
-            let _: miniscript::Miniscript<_, miniscript::Segwitv0> =
-                miniscript::Miniscript::parse(script)
-                    .map_err(|_| PsbtValidationError::InvalidInWitnessScript(input.clone()))?;
-        }
-
-        // We'll then check it doesn't create more than it spends
-        value_in = value_in
-            .checked_add(
-                input
-                    .witness_utxo
-                    .as_ref()
-                    .expect("None checked above")
-                    .value,
-            )
-            .ok_or(PsbtValidationError::InsaneAmounts)?;
-    }
-
-    let mut value_out: u64 = 0;
-    for o in inner_tx.output.iter() {
-        value_out = value_out
-            .checked_add(o.value)
-            .ok_or(PsbtValidationError::InsaneAmounts)?;
-    }
-
-    if value_out > value_in {
-        return Err(PsbtValidationError::InsaneAmounts);
-    }
-
-    Ok(psbt)
-}
-
-fn find_revocationtx_input(inputs: &[PsbtIn]) -> Option<&PsbtIn> {
-    inputs.iter().find(|i| {
-        i.witness_utxo
-            .as_ref()
-            .map(|o| o.script_pubkey.is_v0_p2wsh())
-            == Some(true)
-    })
-}
-
-fn find_feebumping_input(inputs: &[PsbtIn]) -> Option<&PsbtIn> {
-    inputs.iter().find(|i| {
-        i.witness_utxo
-            .as_ref()
-            .map(|o| o.script_pubkey.is_v0_p2wpkh())
-            == Some(true)
-    })
-}
-
-// The Cancel, Emer and Unvault Emer are Revocation transactions
-fn check_revocationtx_input(input: &PsbtIn) -> Result<(), PsbtValidationError> {
-    if input.final_script_witness.is_some() {
-        // Already final, sighash type and witness script are wiped
-        return Ok(());
-    }
-
-    // The revocation input must indicate that it wants to be signed with ACP
-    if input.sighash_type != Some(SigHashType::AllPlusAnyoneCanPay) {
-        return Err(PsbtValidationError::InvalidSighashType(input.clone()));
-    }
-
-    // The revocation input must contain a valid witness script
-    if let Some(ref ws) = input.witness_script {
-        if Some(&ws.to_v0_p2wsh()) != input.witness_utxo.as_ref().map(|w| &w.script_pubkey) {
-            return Err(PsbtValidationError::InvalidInWitnessScript(input.clone()));
-        }
-    } else {
-        return Err(PsbtValidationError::MissingInWitnessScript(input.clone()));
-    }
-
-    Ok(())
-}
-
-// The Cancel, Emer and Unvault Emer are Revocation transactions, this checks the appended input to
-// bump the feerate.
-fn check_feebump_input(input: &PsbtIn) -> Result<(), PsbtValidationError> {
-    if input.final_script_witness.is_some() {
-        // Already final, sighash type and witness script are wiped
-        return Ok(());
-    }
-
-    // The feebump input must indicate that it wants to be signed with ALL
-    if input.sighash_type != Some(SigHashType::All) {
-        return Err(PsbtValidationError::InvalidSighashType(input.clone()));
-    }
-
-    // The feebump input must be P2WPKH
-    if input
-        .witness_utxo
-        .as_ref()
-        .map(|u| u.script_pubkey.is_v0_p2wpkh())
-        != Some(true)
-    {
-        return Err(PsbtValidationError::InvalidPrevoutType(input.clone()));
-    }
-
-    // And therefore must not have a witness script
-    if input.witness_script.is_some() {
-        return Err(PsbtValidationError::InvalidInputField(input.clone()));
-    }
-
-    Ok(())
-}
-
-impl_revault_transaction!(
-    UnvaultTransaction,
-    doc = "The unvaulting transaction, spending a deposit and being eventually spent by a spend transaction (if not revaulted)."
-);
-impl UnvaultTransaction {
-    // Internal DRY routine for creating the inner PSBT
-    fn create_psbt(
-        deposit_txin: DepositTxIn,
-        unvault_txout: UnvaultTxOut,
-        cpfp_txout: CpfpTxOut,
-        lock_time: u32,
-    ) -> Psbt {
-        Psbt {
-            global: PsbtGlobal {
-                unsigned_tx: Transaction {
-                    version: TX_VERSION,
-                    lock_time,
-                    input: vec![deposit_txin.unsigned_txin()],
-                    output: vec![unvault_txout.into_txout(), cpfp_txout.into_txout()],
-                },
-                version: 0,
-                xpub: BTreeMap::new(),
-                proprietary: BTreeMap::new(),
-                unknown: BTreeMap::new(),
-            },
-            inputs: vec![PsbtIn {
-                witness_script: Some(deposit_txin.txout().witness_script().clone()),
-                sighash_type: Some(SigHashType::All),
-                witness_utxo: Some(deposit_txin.into_txout().into_txout()),
-                ..PsbtIn::default()
-            }],
-            // 1 Unvault, 1 CPFP
-            outputs: vec![PsbtOut::default(), PsbtOut::default()],
-        }
-    }
-
-    /// An unvault transaction always spends one deposit output and contains one CPFP output in
-    /// addition to the unvault one.
-    /// It's always created using a fixed feerate and the CPFP output value is fixed as well.
-    ///
-    /// BIP174 Creator and Updater roles.
-    pub fn new(
-        deposit_input: DepositTxIn,
-        unvault_descriptor: &DerivedUnvaultDescriptor,
-        cpfp_descriptor: &DerivedCpfpDescriptor,
-        lock_time: u32,
-    ) -> Result<UnvaultTransaction, TransactionCreationError> {
-        // First, create a dummy transaction to get its weight without Witness
-        let dummy_unvault_txout = UnvaultTxOut::new(u64::MAX, unvault_descriptor);
-        let dummy_cpfp_txout = CpfpTxOut::new(u64::MAX, cpfp_descriptor);
-        let dummy_tx = UnvaultTransaction::create_psbt(
-            deposit_input.clone(),
-            dummy_unvault_txout,
-            dummy_cpfp_txout,
-            lock_time,
-        )
-        .global
-        .unsigned_tx;
-
-        // The weight of the transaction once signed will be the size of the witness-stripped
-        // transaction plus the size of the single input's witness.
-        let total_weight = dummy_tx
-            .get_weight()
-            .checked_add(deposit_input.txout().max_sat_weight())
-            .expect("Properly-computed weights cannot overflow");
-        let total_weight: u64 = total_weight.try_into().expect("usize in u64");
-        let fees = UNVAULT_TX_FEERATE
-            .checked_mul(total_weight)
-            .expect("Properly-computed weights cannot overflow");
-        // Nobody wants to pay 3k€ fees if we had a bug.
-        if fees > INSANE_FEES {
-            return Err(TransactionCreationError::InsaneFees);
-        }
-
-        assert!(
-            total_weight <= MAX_STANDARD_TX_WEIGHT as u64,
-            "A single input and two outputs"
-        );
-
-        // The unvault output value is then equal to the deposit value minus the fees and the CPFP.
-        let deposit_value = deposit_input.txout().txout().value;
-        if fees + UNVAULT_CPFP_VALUE + DUST_LIMIT > deposit_value {
-            return Err(TransactionCreationError::Dust);
-        }
-        let unvault_value = deposit_value - fees - UNVAULT_CPFP_VALUE; // Arithmetic checked above
-
-        let unvault_txout = UnvaultTxOut::new(unvault_value, unvault_descriptor);
-        let cpfp_txout = CpfpTxOut::new(UNVAULT_CPFP_VALUE, cpfp_descriptor);
-        Ok(UnvaultTransaction(UnvaultTransaction::create_psbt(
-            deposit_input,
-            unvault_txout,
-            cpfp_txout,
-            lock_time,
-        )))
-    }
-
-    fn unvault_txin(
-        &self,
-        unvault_descriptor: &DerivedUnvaultDescriptor,
-        sequence: u32,
-    ) -> UnvaultTxIn {
-        let spk = unvault_descriptor.inner().script_pubkey();
-        let index = self
-            .inner_tx()
-            .global
-            .unsigned_tx
-            .output
-            .iter()
-            .position(|txo| txo.script_pubkey == spk)
-            .expect("UnvaultTransaction is always created with an Unvault txo");
-
-        // Unwraped above
-        let txo = &self.inner_tx().global.unsigned_tx.output[index];
-        let prev_txout = UnvaultTxOut::new(txo.value, unvault_descriptor);
-        UnvaultTxIn::new(
-            OutPoint {
-                txid: self.inner_tx().global.unsigned_tx.txid(),
-                vout: index.try_into().expect("There are two outputs"),
-            },
-            prev_txout,
-            sequence,
-        )
-    }
-
-    /// Get the Unvault txo to be referenced in a spending transaction
-    pub fn spend_unvault_txin(&self, unvault_descriptor: &DerivedUnvaultDescriptor) -> UnvaultTxIn {
-        self.unvault_txin(unvault_descriptor, unvault_descriptor.csv_value())
-    }
-
-    /// Get the Unvault txo to be referenced in a revocation transaction
-    pub fn revault_unvault_txin(
-        &self,
-        unvault_descriptor: &DerivedUnvaultDescriptor,
-    ) -> UnvaultTxIn {
-        self.unvault_txin(unvault_descriptor, RBF_SEQUENCE)
-    }
-
-    /// Get the CPFP txo to be referenced in a spending transaction
-    pub fn cpfp_txin(&self, cpfp_descriptor: &DerivedCpfpDescriptor) -> CpfpTxIn {
-        let spk = cpfp_descriptor.inner().script_pubkey();
-        let index = self
-            .inner_tx()
-            .global
-            .unsigned_tx
-            .output
-            .iter()
-            .position(|txo| txo.script_pubkey == spk)
-            .expect("We always create UnvaultTransaction with a CPFP output");
-
-        // Unwraped above
-        let txo = &self.inner_tx().global.unsigned_tx.output[index];
-        let prev_txout = CpfpTxOut::new(txo.value, cpfp_descriptor);
-        CpfpTxIn::new(
-            OutPoint {
-                txid: self.inner_tx().global.unsigned_tx.txid(),
-                vout: index.try_into().expect("There are two outputs"),
-            },
-            prev_txout,
-        )
-    }
-
-    /// Parse an Unvault transaction from a PSBT
-    pub fn from_raw_psbt(raw_psbt: &[u8]) -> Result<Self, TransactionSerialisationError> {
-        let psbt = Decodable::consensus_decode(raw_psbt)?;
-        let psbt = psbt_common_sanity_checks(psbt)?;
-
-        // Unvault + CPFP txos
-        let output_count = psbt.global.unsigned_tx.output.len();
-        if output_count != 2 {
-            return Err(PsbtValidationError::InvalidOutputCount(output_count).into());
-        }
-
-        let input_count = psbt.global.unsigned_tx.input.len();
-        // We for now have 1 unvault == 1 deposit
-        if input_count != 1 {
-            return Err(PsbtValidationError::InvalidInputCount(input_count).into());
-        }
-        let input = &psbt.inputs[0];
-        if input.final_script_witness.is_none() {
-            if input.sighash_type != Some(SigHashType::All) {
-                return Err(PsbtValidationError::InvalidSighashType(input.clone()).into());
-            }
-            if let Some(ref ws) = input.witness_script {
-                if ws.to_v0_p2wsh()
-                    != input
-                        .witness_utxo
-                        .as_ref()
-                        .expect("Check in sanity checks")
-                        .script_pubkey
-                {
-                    return Err(PsbtValidationError::InvalidInWitnessScript(input.clone()).into());
-                }
-            } else {
-                return Err(PsbtValidationError::MissingInWitnessScript(input.clone()).into());
-            }
-        }
-
-        // NOTE: the Unvault transaction cannot get larger than MAX_STANDARD_TX_WEIGHT
-
-        Ok(UnvaultTransaction(psbt))
-    }
-}
-
-impl_revault_transaction!(
-    CancelTransaction,
-    doc = "The transaction \"revaulting\" a spend attempt, i.e. spending the unvaulting transaction back to a deposit txo."
-);
-impl CancelTransaction {
-    // Internal DRY routine for creating the inner PSBT
-    fn create_psbt(
-        unvault_txin: UnvaultTxIn,
-        feebump_txin: Option<FeeBumpTxIn>,
-        deposit_txo: DepositTxOut,
-        lock_time: u32,
-    ) -> Psbt {
-        let mut txins = vec![unvault_txin.unsigned_txin()];
-        let mut psbtins = vec![PsbtIn {
-            witness_script: Some(unvault_txin.txout().witness_script().clone()),
-            sighash_type: Some(SigHashType::AllPlusAnyoneCanPay),
-            witness_utxo: Some(unvault_txin.into_txout().into_txout()),
-            ..PsbtIn::default()
-        }];
-        if let Some(feebump_txin) = feebump_txin {
-            txins.push(feebump_txin.unsigned_txin());
-            psbtins.push(PsbtIn {
-                sighash_type: Some(SigHashType::All),
-                witness_utxo: Some(feebump_txin.into_txout().into_txout()),
-                ..PsbtIn::default()
-            });
-        }
-
-        Psbt {
-            global: PsbtGlobal {
-                unsigned_tx: Transaction {
-                    version: TX_VERSION,
-                    lock_time,
-                    input: txins,
-                    output: vec![deposit_txo.into_txout()],
-                },
-                version: 0,
-                xpub: BTreeMap::new(),
-                proprietary: BTreeMap::new(),
-                unknown: BTreeMap::new(),
-            },
-            inputs: psbtins,
-            // Deposit txout
-            outputs: vec![PsbtOut::default()],
-        }
-    }
-
-    /// A cancel transaction always pays to a deposit output and spends the unvault output, and
-    /// may have a fee-bumping input.
-    ///
-    /// BIP174 Creator and Updater roles.
-    pub fn new(
-        unvault_input: UnvaultTxIn,
-        feebump_input: Option<FeeBumpTxIn>,
-        deposit_descriptor: &DerivedDepositDescriptor,
-        lock_time: u32,
-    ) -> CancelTransaction {
-        // First, create a dummy transaction to get its weight without Witness. Note that we always
-        // account for the weight *without* feebump input. It pays for itself.
-        let dummy_deposit_txo = DepositTxOut::new(u64::MAX, deposit_descriptor);
-        let dummy_tx = CancelTransaction::create_psbt(
-            unvault_input.clone(),
-            None,
-            dummy_deposit_txo,
-            lock_time,
-        )
-        .global
-        .unsigned_tx;
-
-        // The weight of the cancel transaction without a feebump input is the weight of the
-        // witness-stripped transaction plus the weight required to satisfy the unvault txin
-        let total_weight = dummy_tx
-            .get_weight()
-            .checked_add(unvault_input.txout().max_sat_weight())
-            .expect("Properly computed weight won't overflow");
-        let total_weight: u64 = total_weight.try_into().expect("usize in u64");
-        let fees = REVAULTING_TX_FEERATE
-            .checked_mul(total_weight)
-            .expect("Properly computed weight won't overflow");
-        // Without the feebump input, it should not be reachable.
-        debug_assert!(fees < INSANE_FEES);
-
-        assert!(
-            total_weight <= MAX_STANDARD_TX_WEIGHT as u64,
-            "At most 2 inputs and single output"
-        );
-
-        // Now, get the revaulting output value out of it.
-        let unvault_value = unvault_input.txout().txout().value;
-        let revault_value = unvault_value
-            .checked_sub(fees)
-            .expect("We would not create a dust unvault txo");
-        let deposit_txo = DepositTxOut::new(revault_value, deposit_descriptor);
-
-        CancelTransaction(CancelTransaction::create_psbt(
-            unvault_input,
-            feebump_input,
-            deposit_txo,
-            lock_time,
-        ))
-    }
-
-    /// Parse a Cancel transaction from a PSBT
-    pub fn from_raw_psbt(raw_psbt: &[u8]) -> Result<Self, TransactionSerialisationError> {
-        let psbt = Decodable::consensus_decode(raw_psbt)?;
-        let psbt = psbt_common_sanity_checks(psbt)?;
-
-        // Deposit txo
-        let output_count = psbt.global.unsigned_tx.output.len();
-        if output_count != 1 {
-            return Err(PsbtValidationError::InvalidOutputCount(output_count).into());
-        }
-
-        // Deposit txo is P2WSH
-        let output = &psbt.outputs[0];
-        if output.witness_script.is_none() {
-            return Err(PsbtValidationError::MissingOutWitnessScript(output.clone()).into());
-        }
-        if output.redeem_script.is_some() {
-            return Err(PsbtValidationError::InvalidOutputField(output.clone()).into());
-        }
-
-        let input_count = psbt.global.unsigned_tx.input.len();
-        if input_count > 2 {
-            return Err(PsbtValidationError::InvalidInputCount(input_count).into());
-        }
-        if input_count > 1 {
-            let input = find_feebumping_input(&psbt.inputs)
-                .ok_or(PsbtValidationError::MissingFeeBumpingInput)?;
-            check_feebump_input(&input)?;
-        }
-        let input = find_revocationtx_input(&psbt.inputs)
-            .ok_or(PsbtValidationError::MissingRevocationInput)?;
-        check_revocationtx_input(&input)?;
-
-        Ok(CancelTransaction(psbt))
-    }
-}
-
-impl_revault_transaction!(
-    EmergencyTransaction,
-    doc = "The transaction spending a deposit output to The Emergency Script."
-);
-impl EmergencyTransaction {
-    // Internal DRY routine for creating the inner PSBT
-    fn create_psbt(
-        deposit_txin: DepositTxIn,
-        feebump_txin: Option<FeeBumpTxIn>,
-        emergency_txo: EmergencyTxOut,
-        lock_time: u32,
-    ) -> Psbt {
-        let mut txins = vec![deposit_txin.unsigned_txin()];
-        let mut psbtins = vec![PsbtIn {
-            witness_script: Some(deposit_txin.txout().witness_script().clone()),
-            sighash_type: Some(SigHashType::AllPlusAnyoneCanPay),
-            witness_utxo: Some(deposit_txin.into_txout().into_txout()),
-            ..PsbtIn::default()
-        }];
-        if let Some(feebump_txin) = feebump_txin {
-            txins.push(feebump_txin.unsigned_txin());
-            psbtins.push(PsbtIn {
-                sighash_type: Some(SigHashType::All),
-                witness_utxo: Some(feebump_txin.into_txout().into_txout()),
-                ..PsbtIn::default()
-            });
-        }
-
-        Psbt {
-            global: PsbtGlobal {
-                unsigned_tx: Transaction {
-                    version: TX_VERSION,
-                    lock_time,
-                    input: txins,
-                    output: vec![emergency_txo.into_txout()],
-                },
-                version: 0,
-                xpub: BTreeMap::new(),
-                proprietary: BTreeMap::new(),
-                unknown: BTreeMap::new(),
-            },
-            inputs: psbtins,
-            // Deposit txout
-            outputs: vec![PsbtOut::default()],
-        }
-    }
-
-    /// The first emergency transaction always spends a deposit output and pays to the Emergency
-    /// Script. It may also spend an additional output for fee-bumping.
-    /// Will error **only** when trying to spend a dust deposit.
-    ///
-    /// BIP174 Creator and Updater roles.
-    pub fn new(
-        deposit_input: DepositTxIn,
-        feebump_input: Option<FeeBumpTxIn>,
-        emer_address: EmergencyAddress,
-        lock_time: u32,
-    ) -> Result<EmergencyTransaction, TransactionCreationError> {
-        // First, create a dummy transaction to get its weight without Witness. Note that we always
-        // account for the weight *without* feebump input. It has to pay for itself.
-        let emer_txo = EmergencyTxOut::new(emer_address.clone(), u64::MAX);
-        let dummy_tx =
-            EmergencyTransaction::create_psbt(deposit_input.clone(), None, emer_txo, lock_time)
-                .global
-                .unsigned_tx;
-
-        // The weight of the emergency transaction without a feebump input is the weight of the
-        // witness-stripped transaction plus the weight required to satisfy the deposit txin
-        let total_weight = dummy_tx
-            .get_weight()
-            .checked_add(deposit_input.txout().max_sat_weight())
-            .expect("Weight computation bug");
-        let total_weight: u64 = total_weight.try_into().expect("usize in u64");
-        let fees = REVAULTING_TX_FEERATE
-            .checked_mul(total_weight)
-            .expect("Weight computation bug");
-        // Without the feebump input, it should not be reachable.
-        debug_assert!(fees < INSANE_FEES);
-
-        assert!(
-            total_weight <= MAX_STANDARD_TX_WEIGHT as u64,
-            "At most 2 inputs and a single output"
-        );
-
-        // Now, get the emergency output value out of it.
-        let deposit_value = deposit_input.txout().txout().value;
-        let emer_value = deposit_value
-            .checked_sub(fees)
-            .ok_or_else(|| TransactionCreationError::Dust)?;
-        let emer_txo = EmergencyTxOut::new(emer_address, emer_value);
-
-        Ok(EmergencyTransaction(EmergencyTransaction::create_psbt(
-            deposit_input.clone(),
-            feebump_input,
-            emer_txo,
-            lock_time,
-        )))
-    }
-
-    /// Parse an Emergency transaction from a PSBT
-    pub fn from_raw_psbt(raw_psbt: &[u8]) -> Result<Self, TransactionSerialisationError> {
-        let psbt = Decodable::consensus_decode(raw_psbt)?;
-        let psbt = psbt_common_sanity_checks(psbt)?;
-
-        // Emergency txo
-        let output_count = psbt.global.unsigned_tx.output.len();
-        if output_count != 1 {
-            return Err(PsbtValidationError::InvalidOutputCount(output_count).into());
-        }
-
-        let input_count = psbt.global.unsigned_tx.input.len();
-        if input_count > 2 {
-            return Err(PsbtValidationError::InvalidInputCount(input_count).into());
-        }
-        if input_count > 1 {
-            let input = find_feebumping_input(&psbt.inputs)
-                .ok_or(PsbtValidationError::MissingFeeBumpingInput)?;
-            check_feebump_input(&input)?;
-        }
-        let input = find_revocationtx_input(&psbt.inputs)
-            .ok_or(PsbtValidationError::MissingRevocationInput)?;
-        check_revocationtx_input(&input)?;
-
-        Ok(EmergencyTransaction(psbt))
-    }
-}
-
-impl_revault_transaction!(
-    UnvaultEmergencyTransaction,
-    doc = "The transaction spending an unvault output to The Emergency Script."
-);
-impl UnvaultEmergencyTransaction {
-    // Internal DRY routine for creating the inner PSBT
-    fn create_psbt(
-        unvault_txin: UnvaultTxIn,
-        feebump_txin: Option<FeeBumpTxIn>,
-        emergency_txo: EmergencyTxOut,
-        lock_time: u32,
-    ) -> Psbt {
-        let mut txins = vec![unvault_txin.unsigned_txin()];
-        let mut psbtins = vec![PsbtIn {
-            witness_script: Some(unvault_txin.txout().witness_script().clone()),
-            sighash_type: Some(SigHashType::AllPlusAnyoneCanPay),
-            witness_utxo: Some(unvault_txin.into_txout().into_txout()),
-            ..PsbtIn::default()
-        }];
-        if let Some(feebump_txin) = feebump_txin {
-            txins.push(feebump_txin.unsigned_txin());
-            psbtins.push(PsbtIn {
-                sighash_type: Some(SigHashType::All),
-                witness_utxo: Some(feebump_txin.into_txout().into_txout()),
-                ..PsbtIn::default()
-            });
-        }
-
-        Psbt {
-            global: PsbtGlobal {
-                unsigned_tx: Transaction {
-                    version: TX_VERSION,
-                    lock_time,
-                    input: txins,
-                    output: vec![emergency_txo.into_txout()],
-                },
-                version: 0,
-                xpub: BTreeMap::new(),
-                proprietary: BTreeMap::new(),
-                unknown: BTreeMap::new(),
-            },
-            inputs: psbtins,
-            // Deposit txout
-            outputs: vec![PsbtOut::default()],
-        }
-    }
-
-    /// The second emergency transaction always spends an unvault output and pays to the Emergency
-    /// Script. It may also spend an additional output for fee-bumping.
-    ///
-    /// BIP174 Creator and Updater roles.
-    pub fn new(
-        unvault_input: UnvaultTxIn,
-        feebump_input: Option<FeeBumpTxIn>,
-        emer_address: EmergencyAddress,
-        lock_time: u32,
-    ) -> UnvaultEmergencyTransaction {
-        // First, create a dummy transaction to get its weight without Witness. Note that we always
-        // account for the weight *without* feebump input. It has to pay for itself.
-        let emer_txo = EmergencyTxOut::new(emer_address.clone(), u64::MAX);
-        let dummy_tx = UnvaultEmergencyTransaction::create_psbt(
-            unvault_input.clone(),
-            None,
-            emer_txo,
-            lock_time,
-        )
-        .global
-        .unsigned_tx;
-
-        // The weight of the unvault emergency transaction without a feebump input is the weight of
-        // the witness-stripped transaction plus the weight required to satisfy the unvault txin
-        let total_weight = dummy_tx
-            .get_weight()
-            .checked_add(unvault_input.txout().max_sat_weight())
-            .expect("Weight computation bug");
-        let total_weight: u64 = total_weight.try_into().expect("usize in u64");
-        let fees = REVAULTING_TX_FEERATE
-            .checked_mul(total_weight)
-            .expect("Weight computation bug");
-        // Without the feebump input, it should not be reachable.
-        debug_assert!(fees < INSANE_FEES);
-
-        assert!(
-            total_weight <= MAX_STANDARD_TX_WEIGHT as u64,
-            "At most 2 inputs and a single output"
-        );
-
-        // Now, get the emergency output value out of it.
-        let deposit_value = unvault_input.txout().txout().value;
-        let emer_value = deposit_value
-            .checked_sub(fees)
-            .expect("We would never create a dust unvault txo");
-        let emer_txo = EmergencyTxOut::new(emer_address, emer_value);
-
-        UnvaultEmergencyTransaction(UnvaultEmergencyTransaction::create_psbt(
-            unvault_input.clone(),
-            feebump_input,
-            emer_txo,
-            lock_time,
-        ))
-    }
-
-    /// Parse an UnvaultEmergency transaction from a PSBT
-    pub fn from_raw_psbt(raw_psbt: &[u8]) -> Result<Self, TransactionSerialisationError> {
-        let psbt = Decodable::consensus_decode(raw_psbt)?;
-        let psbt = psbt_common_sanity_checks(psbt)?;
-
-        // Emergency txo
-        let output_count = psbt.global.unsigned_tx.output.len();
-        if output_count != 1 {
-            return Err(PsbtValidationError::InvalidOutputCount(output_count).into());
-        }
-
-        let input_count = psbt.global.unsigned_tx.input.len();
-        if input_count > 2 {
-            return Err(PsbtValidationError::InvalidInputCount(input_count).into());
-        }
-        if input_count > 1 {
-            let input = find_feebumping_input(&psbt.inputs)
-                .ok_or(PsbtValidationError::MissingFeeBumpingInput)?;
-            check_feebump_input(&input)?;
-        }
-        let input = find_revocationtx_input(&psbt.inputs)
-            .ok_or(PsbtValidationError::MissingRevocationInput)?;
-        check_revocationtx_input(&input)?;
-
-        Ok(UnvaultEmergencyTransaction(psbt))
-    }
-}
-
-impl_revault_transaction!(
-    SpendTransaction,
-    doc = "The transaction spending the unvaulting transaction, paying to one or multiple \
-    externally-controlled addresses, and possibly to a new deposit txo for the change."
-);
-impl SpendTransaction {
-    /// A spend transaction can batch multiple unvault txouts, and may have any number of
-    /// txouts (destination and change) in addition to the CPFP one..
-    ///
-    /// The insane fees check is gated behind the `insane_fee_checks` parameter as the caller
-    /// may want to create a transaction without a change output.
-    ///
-    /// BIP174 Creator and Updater roles.
-    pub fn new(
-        unvault_inputs: Vec<UnvaultTxIn>,
-        spend_txouts: Vec<SpendTxOut>,
-        cpfp_descriptor: &DerivedCpfpDescriptor,
-        lock_time: u32,
-        insane_fee_check: bool,
-    ) -> Result<SpendTransaction, TransactionCreationError> {
-        // The CPFP is tricky to compute. We could be smart and avoid some allocations here
-        // but at the cost of clarity.
-        let cpfp_txo = SpendTransaction::cpfp_txout(
-            unvault_inputs.clone(),
-            spend_txouts.clone(),
-            cpfp_descriptor,
-            lock_time,
-        );
-
-        // Used later to check the maximum transaction size.
-        let sat_weight = unvault_inputs
-            .iter()
-            .map(|txin| txin.txout().max_sat_weight())
-            .sum::<usize>();
-
-        // Record the value spent
-        let mut value_in: u64 = 0;
-
-        let mut txos = Vec::with_capacity(spend_txouts.len() + 1);
-        txos.push(cpfp_txo.txout().clone());
-        txos.extend(spend_txouts.iter().map(|spend_txout| match spend_txout {
-            SpendTxOut::Destination(ref txo) => txo.clone(),
-            SpendTxOut::Change(ref txo) => txo.clone().into_txout(),
-        }));
-        let psbtouts = txos.iter().map(|_| PsbtOut::default()).collect();
-
-        let psbt = Psbt {
-            global: PsbtGlobal {
-                unsigned_tx: Transaction {
-                    version: TX_VERSION,
-                    lock_time,
-                    input: unvault_inputs
-                        .iter()
-                        .map(|input| input.unsigned_txin())
-                        .collect(),
-                    output: txos,
-                },
-                version: 0,
-                xpub: BTreeMap::new(),
-                proprietary: BTreeMap::new(),
-                unknown: BTreeMap::new(),
-            },
-            inputs: unvault_inputs
-                .into_iter()
-                .map(|input| {
-                    let prev_txout = input.into_txout();
-                    value_in += prev_txout.txout().value;
-                    PsbtIn {
-                        witness_script: Some(prev_txout.witness_script().clone()),
-                        sighash_type: Some(SigHashType::All), // Unvault spends are always signed with ALL
-                        witness_utxo: Some(prev_txout.into_txout()),
-                        ..PsbtIn::default()
-                    }
-                })
-                .collect(),
-            outputs: psbtouts,
-        };
-
-        // Make sure we didn't create a Monster Tx :tm: ..
-        let unsigned_tx = &psbt.global.unsigned_tx;
-        let witstrip_weight = unsigned_tx.get_weight();
-        let total_weight = sat_weight
-            .checked_add(witstrip_weight)
-            .expect("Weight computation bug: cannot overflow");
-        if total_weight > MAX_STANDARD_TX_WEIGHT as usize {
-            return Err(TransactionCreationError::TooLarge);
-        }
-
-        let value_out: u64 = unsigned_tx.output.iter().map(|o| o.value).sum();
-        let fees = value_in
-            .checked_sub(value_out)
-            .ok_or(TransactionCreationError::NegativeFees)?;
-        if insane_fee_check && fees > INSANE_FEES {
-            return Err(TransactionCreationError::InsaneFees);
-        }
-
-        Ok(SpendTransaction(psbt))
-    }
-
-    /// Get the CPFP transaction output for a Spend transaction spending these `unvault_inputs`
-    /// and creating these `spend_txouts`.
-    ///
-    /// The CPFP output value is dependant on the transaction size, see [practical-revaul
-    /// t](https://github.com/revault/practical-revault/blob/master/transactions.md#spend_tx) for
-    /// more details.
-    pub fn cpfp_txout(
-        unvault_inputs: Vec<UnvaultTxIn>,
-        spend_txouts: Vec<SpendTxOut>,
-        cpfp_descriptor: &DerivedCpfpDescriptor,
-        lock_time: u32,
-    ) -> CpfpTxOut {
-        let mut txos = Vec::with_capacity(spend_txouts.len() + 1);
-        let dummy_cpfp_txo = CpfpTxOut::new(u64::MAX, &cpfp_descriptor);
-        txos.push(dummy_cpfp_txo.txout().clone());
-        txos.extend(spend_txouts.iter().map(|spend_txout| match spend_txout {
-            SpendTxOut::Destination(ref txo) => txo.clone(),
-            SpendTxOut::Change(ref txo) => txo.clone().into_txout(),
-        }));
-        let dummy_tx = Transaction {
-            version: TX_VERSION,
-            lock_time,
-            input: unvault_inputs
-                .iter()
-                .map(|input| input.unsigned_txin())
-                .collect(),
-            output: txos,
-        };
-
-        let sat_weight: u64 = unvault_inputs
-            .iter()
-            .map(|txin| txin.txout().max_sat_weight())
-            .sum::<usize>()
-            .try_into()
-            .expect("An usize doesn't fit in an u64?");
-        let witstrip_weight: u64 = dummy_tx
-            .get_weight()
-            .try_into()
-            .expect("Bug: an usize that doesn't fit in a u64?");
-        let total_weight = sat_weight
-            .checked_add(witstrip_weight)
-            .expect("Weight computation bug: cannot overflow");
-
-        // See https://github.com/revault/practical-revault/blob/master/transactions.md#spend_tx
-        // for this arbirtrary value.
-        let cpfp_value = 16 * total_weight;
-        CpfpTxOut::new(cpfp_value, &cpfp_descriptor)
-    }
-
-    /// Get the feerate of this transaction, assuming fully-satisfied inputs. If the transaction
-    /// is already finalized, returns the exact feerate. Otherwise computes the maximum reasonable
-    /// weight of a satisfaction and returns the feerate based on this estimation.
-    pub fn max_feerate(&self) -> u64 {
-        let fees = self.fees();
-        let weight = self.max_weight();
-
-        fees.checked_add(weight - 1) // Weight is never 0
-            .expect("Feerate computation bug, fees >u64::MAX")
-            .checked_div(weight)
-            .expect("Weight is never 0")
-    }
-
-    /// Get the size of this transaction, assuming fully-satisfied inputs. If the transaction
-    /// is already finalized, returns the exact size in witness units. Otherwise computes the
-    /// maximum reasonable weight of a satisfaction.
-    pub fn max_weight(&self) -> u64 {
-        let psbt = self.inner_tx();
-        let tx = &psbt.global.unsigned_tx;
-
-        let mut weight: u64 = tx.get_weight().try_into().expect("Can't be >u64::MAX");
-        for txin in psbt.inputs.iter() {
-            let txin_weight: u64 = if self.is_finalized() {
-                txin.final_script_witness
-                    .as_ref()
-                    .expect("Always set if final")
-                    .iter()
-                    .map(|e| e.len())
-                    .sum::<usize>()
-                    .try_into()
-                    .expect("Bug: witness size >u64::MAX")
-            } else {
-                // FIXME: this panic can probably be triggered...
-                miniscript::descriptor::Wsh::new(
-                    miniscript::Miniscript::parse(
-                        txin.witness_script
-                            .as_ref()
-                            .expect("Unvault txins always have a witness Script"),
-                    )
-                    .expect("UnvaultTxIn witness_script is created from a Miniscript"),
-                )
-                .expect("")
-                .max_satisfaction_weight()
-                .expect("It's a sane Script, derived from a Miniscript")
-                .try_into()
-                .expect("Can't be >u64::MAX")
-            };
-            weight = weight
-                .checked_add(txin_weight)
-                .expect("Weight computation bug: overflow computing spent coins value");
-        }
-        assert!(weight > 0, "We never create an empty tx");
-
-        weight
-    }
-
-    // FIXME: feerate sanity checks
-    /// Parse a Spend transaction from a PSBT
-    pub fn from_raw_psbt(raw_psbt: &[u8]) -> Result<Self, TransactionSerialisationError> {
-        let psbt = Decodable::consensus_decode(raw_psbt)?;
-        let psbt = psbt_common_sanity_checks(psbt)?;
-
-        if psbt.inputs.is_empty() {
-            return Err(PsbtValidationError::InvalidInputCount(0).into());
-        }
-
-        let mut max_sat_weight = 0;
-        for input in psbt.inputs.iter() {
-            if input.final_script_witness.is_some() {
-                continue;
-            }
-
-            if input.sighash_type != Some(SigHashType::All) {
-                return Err(PsbtValidationError::InvalidSighashType(input.clone()).into());
-            }
-
-            // The revocation input must contain a valid witness script
-            if let Some(ref ws) = input.witness_script {
-                if Some(&ws.to_v0_p2wsh()) != input.witness_utxo.as_ref().map(|w| &w.script_pubkey)
-                {
-                    return Err(PsbtValidationError::InvalidInWitnessScript(input.clone()).into());
-                }
-            } else {
-                return Err(PsbtValidationError::MissingInWitnessScript(input.clone()).into());
-            }
-
-            max_sat_weight += miniscript::descriptor::Wsh::new(
-                miniscript::Miniscript::parse(
-                    input
-                        .witness_script
-                        .as_ref()
-                        .ok_or_else(|| PsbtValidationError::InvalidInputField(input.clone()))?,
-                )
-                .map_err(|_| PsbtValidationError::InvalidInputField(input.clone()))?,
-            )
-            .map_err(|_| PsbtValidationError::InvalidInputField(input.clone()))?
-            .max_satisfaction_weight()
-            .map_err(|_| PsbtValidationError::InvalidInputField(input.clone()))?;
-        }
-
-        // Make sure the transaction cannot get out of standardness bounds once finalized
-        let spend_tx = SpendTransaction(psbt);
-        let witstrip_weight = spend_tx.inner_tx().global.unsigned_tx.get_weight();
-        let total_weight = witstrip_weight
-            .checked_add(max_sat_weight)
-            .expect("Weight computation bug");
-        if total_weight > MAX_STANDARD_TX_WEIGHT as usize {
-            return Err(PsbtValidationError::TransactionTooLarge.into());
-        }
-
-        Ok(spend_tx)
     }
 }
 
@@ -2336,6 +1250,7 @@ mod tests {
                 };
             }
 
+            roundtrip!(cancel_tx);
             roundtrip!(emergency_tx);
             roundtrip!(unvault_tx);
             roundtrip!(unemergency_tx);
